# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FreeAgentics API
  description: |
    The FreeAgentics API provides a comprehensive interface for managing multi-agent AI systems with Active Inference capabilities. 
    
    This API allows you to:
    - Create and manage AI agents with different templates
    - Run inference tasks with context-aware processing
    - Build and query knowledge graphs
    - Monitor system performance and agent metrics
    - Implement real-time communication via WebSockets
    
    ## Authentication
    
    All API endpoints require authentication using JWT tokens. Obtain tokens by registering or logging in, then include the access token in the Authorization header:
    
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    
    The API implements rate limiting to ensure fair usage:
    - Auth endpoints: 3 requests per minute
    - API endpoints: 60 requests per minute
    - WebSocket connections: 100 messages per minute
    
    Rate limit headers are included in all responses.
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns structured error responses:
    
    ```json
    {
      "error": {
        "code": "INVALID_REQUEST",
        "message": "The request was invalid",
        "details": {
          "field": "name",
          "reason": "Required field missing"
        }
      },
      "request_id": "req_12345"
    }
    ```
  version: 0.1.0
  contact:
    name: FreeAgentics Support
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: Retrieve a list of agents accessible to the user
      operationId: listAgents
      parameters:
        - name: skip
          in: query
          description: Number of agents to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of agents to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: status
          in: query
          description: Filter by agent status
          required: false
          schema:
            type: string
            enum: [pending, active, paused, stopped]
        - name: template
          in: query
          description: Filter by agent template
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Agents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Agents
      summary: Create agent
      description: Create a new AI agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentConfig'
            examples:
              research_agent:
                summary: Research agent
                value:
                  name: "Research Assistant"
                  template: "research_v2"
                  parameters:
                    temperature: 0.7
                    max_tokens: 2048
                  use_pymdp: true
                  planning_horizon: 3
      responses:
        '200':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Invalid agent configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or registration
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, RATE_LIMITED, INTERNAL_ERROR]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        request_id:
          type: string
          description: Unique request identifier for support
...
    AgentConfig:
      type: object
      required:
        - name
        - template
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        template:
          type: string
          description: Agent template ID
        parameters:
          type: object
          additionalProperties: true
          default: {}
        gmn_spec:
          type: string
          description: GMN specification for Active Inference model
        use_pymdp:
          type: boolean
          default: true
          description: Whether to use PyMDP for Active Inference
        planning_horizon:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
          description: Planning horizon for Active Inference
...
    Agent:
      type: object
      required:
        - id
        - name
        - template
        - status
        - created_at
        - inference_count
        - parameters
      properties:
        id:
          type: string
        name:
          type: string
        template:
          type: string
        status:
          type: string
          enum: [pending, active, paused, stopped]
        created_at:
          type: string
          format: date-time
        last_active:
          type: string
          format: date-time
        inference_count:
          type: integer
          minimum: 0
        parameters:
          type: object
          additionalProperties: true
    AgentListResponse:
      type: object
      required:
        - agents
        - total
        - skip
        - limit
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
          minimum: 0
        skip:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
...